Remove the broken initialization of the Vault URL in api.NewClient()
The variable c.Address is set with a default address (https://127.0.0.1:8200)
or with the environment variable VAULT_ADDR when set.

Remove the initialization of the Vault URL in api.NewClient() that ends up
in assigning the default address if the VAULT_ADDR is set instead of the
value of the environment variable.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>